requirements.txt
certifi==2021.10.8
Flask==2.0.3
Flask_RESTful==0.3.10
line_bot_sdk==3.1.0
pymongo==3.12.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
pytz==2022.7.1
